% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.heaRt_fit.R
\name{plot.heaRt_fit}
\alias{plot.heaRt_fit}
\title{Plot heaRt output}
\usage{
\method{plot}{heaRt_fit}(x, ...)
}
\arguments{
\item{x}{An abject of class \code{"heaRt_fit"}.}

\item{...}{Catches unused arguments to \code{plot}}
}
\value{
Either a ggplot, a prp plot or partial plot
}
\description{
Plot the different type of model from symptoms related to the diagnosis of heart disease
}
