% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaRt-package.R
\docType{package}
\name{heaRt-package}
\alias{heaRt}
\alias{heaRt-package}
\title{heaRt: Heart Disease Modelling and Visualisation}
\description{
The \pkg{heaRt} package provides tools for loading, analysing and visualising
clinical data related to heart disease diagnoses.
}
\section{Main functions}{
?
\describe{
  \item{\code{\link{load_heaRt}}}{Load and preprocess heart disease clinical data}
  \item{\code{\link{fit}}}{Fit statistical and machine learning models}
  \item{\code{\link{plot.heaRt_fit}}}{Visualise fitted models}
}
}

\section{Data source}{

The data are obtained from the UCI Machine Learning Repository
(Cleveland Heart Disease dataset).
}

\seealso{
\code{\link{load_heaRt}}, \code{\link{fit}}, \code{\link{plot.heaRt_fit}}
}
\author{
\strong{Maintainer}: Nicola Ruaro \email{nicolai.ruaro@gmail.com}

Authors:
\itemize{
  \item Lucrezia Bonifazi \email{lucrezia.bonifazi.2026@mumail.ie}
  \item Anna Signorini \email{anna.signorini.2026@mumail.ie}
}

}
