% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heart.R
\name{load_heaRt}
\alias{load_heaRt}
\title{Load heart disease data from clinical examinations}
\usage{
load_heaRt(
  vars = c("a-s", "rbp-restECG", "chol-mhr"),
  severe_diag = FALSE
)
}
\description{
Downloads and preprocesses heart disease data from the UCI Machine Learning
Repository. The function performs basic data cleaning, constructs a binary
diagnosis variable, and returns a structured object suitable for further
modelling with the \code{\link{fit}} function.
}
\details{
Loads medical data derived from clinical examinations performed on patients,
with the aim of analysing the presence or absence of heart disease.
The function allows the user to select different pairs of clinical
predictors and to focus on either mild or severe diagnoses.
}
\note{
A dedicated \code{\link{fit}} method is provided for objects of class
\code{"heaRt"}.
}
\examples{
dat <- load_heaRt(vars = "rbp-restECG")
dat_severe <- load_heaRt(vars = "chol-mhr", severe_diag = TRUE)

}
\seealso{
\code{\link{fit}}, \code{\link{plot.heaRt_fit}}
}
