% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_heart.R
\name{load_heaRt}
\alias{load_heaRt}
\title{Load in medical data about the diagnosis of a heart disease}
\usage{
load_heaRt(vars = c("a-s", "rbp-restECG", "chol-mhr"), severe_diag = FALSE)
}
\arguments{
\item{vars}{Either \code{"a-s"}, \code{"rbp-restECG"} or \code{"chol-mhr"} for age(1=M, 0=F)-sex,
resting blood pressure-resting ECG and serum cholesterol-maximum heart rate.}

\item{severe_diag}{Either \code{"TRUE"} or \code{"FALSE"}, to select the severe
diagnosis or the normal diagnosis. Defaults to \code{"FALSE"}.}
}
\value{
An object of class \code{"heaRt"} which is a list of the chosen variables and the results
of the diagnosis (1 = positive, 0 = negative)
}
\description{
Load in medical data about medical exams done by patients in order to find out if they have an heart
disease and if it's severe or no.
}
\note{
A dedicated \code{\link{fit}} function is provided for objects of class \code{"heaRt"}.
}
\examples{
dat <- load_heaRt(vars = "rbp-restECG", severe_diag = TRUE)
}
\seealso{
\code{\link{fit}}, \code{\link{plot.heaRt_fit}}
}
